# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ansa dashboard API",
    "version": "v1",
    "description": "Self-service for a tenant organisation: its people, its agent's configuration, and its call history. Every request is scoped to one organisation by the session token that made it."
  },
...
  "paths": {
    "/api/v1/auth/me": {
      "get": {
        "operationId": "auth.me",
        "tags": [
          "auth"
        ],
        "summary": "The signed-in user, their organisation, and what they may do in it",
        "x-ansa-capability": "authenticated",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "maxLength": 320,
                          "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$"
                        },
                        "displayName": {
                          "type": "string",
                          "maxLength": 200
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName"
                      ],
                      "additionalProperties": false
                    },
                    "organisation": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
                        },
                        "name": {
                          "type": "string",
                          "maxLength": 200
                        }
                      },
                      "required": [
                        "id",
                        "name"
                      ],
                      "additionalProperties": false
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "member"
                      ]
                    },
                    "capabilities": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "calls:read",
                          "members:read",
                          "members:write",
                          "invitations:read",
                          "invitations:write",
                          "config:read",
                          "config:write"
                        ]
                      }
                    }
                  },
                  "required": [
                    "user",
                    "organisation",
                    "role",
                    "capabilities"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "No session, or a session that is no longer valid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/organisations": {
      "post": {
        "operationId": "auth.organisations",
        "tags": [
          "auth"
        ],
        "summary": "List the organisations an email and password can sign in to",
        "description": "Returns an empty list for a wrong password and for an address with no account, and takes the same time to do it.",
        "x-ansa-capability": "public",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 320,
                    "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 12,
                    "maxLength": 256
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "organisations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
                          },
                          "name": {
                            "type": "string",
                            "maxLength": 200
                          },
                          "role": {
                            "type": "string",
                            "enum": [
                              "owner",
                              "admin",
                              "member"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "role"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "organisations"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "The request did not match this endpoint's schema",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many attempts; see the Retry-After header",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sessions": {
      "post": {
        "operationId": "auth.signIn",
        "tags": [
          "auth"
        ],
        "summary": "Sign in to one organisation",
        "x-ansa-capability": "public",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 320,
                    "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 12,
                    "maxLength": 256
                  },
                  "organisationId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "organisationId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "organisation": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
                        },
                        "name": {
                          "type": "string",
                          "maxLength": 200
                        }
                      },
                      "required": [
                        "id",
                        "name"
                      ],
                      "additionalProperties": false
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "member"
                      ]
                    }
                  },
                  "required": [
                    "token",
                    "expiresAt",
                    "organisation",
                    "role"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "The request did not match this endpoint's schema",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many attempts; see the Retry-After header",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sessions/current": {
      "delete": {
        "operationId": "auth.signOut",
        "tags": [
          "auth"
        ],
        "summary": "Sign out, revoking the token that made this request",
        "description": "Idempotent. The session row is kept and marked revoked, so it stays in the audit trail.",
        "x-ansa-capability": "authenticated",
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "No session, or a session that is no longer valid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer",
        "description": "The token from POST /api/v1/auth/sessions. It names the organisation it belongs to, so no other header selects one."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457. Every error from this API has this shape.",
        "properties": {
          "type": {
            "type": "string",
            "description": "Stable identifier, prefixed `urn:ansa:problem:`"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "description": "Echoed in the X-Request-Id header"
          },
          "errors": {
            "type": "array",
            "description": "Present on 422 only, one entry per field that did not validate",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "additionalProperties": false
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ],
...
}